optimize and fix bug.

+++ post_process.py
@@ -14,8 +14,8 @@ def process(out, ct, lct, mask, min_v=-1024, max_v=3000):
     out = np.clip(out, min_v, max_v)
     ct = np.clip(ct, min_v, max_v)
 
-    y_min, x_min = int(lct[1]), int(lct[0])
-    y_max, x_max = y_min + int(lct[3]), x_min + int(lct[2])
+    y_min, x_min = lct[1].astype(int), lct[0].astype(int)
+    y_max, x_max = y_min + lct[3].astype(int), x_min + lct[2].astype(int)
 
     # 裁剪出图像区域
     out = out[y_min:y_max, x_min:x_max]
